+++ src/main/resources/application.yml
@@ -33,12 +33,16 @@ spring:
       sort-properties-alphabetically: true
       default-view-inclusion: false
 
-security:
-  user:
-    password: password
+azure:
+  activedirectory:
+    client-secret: t8lFHodtiXdZMU0p4P0SpkLVIEfhg8bkt35DLihC3DQ=
+    client-id: d08628b7-a0cd-4679-bdd7-f12d92f3cb0f
+    active-directory-groups: Editors,EDITOR
 
 server:
-  contextPath: /api
+  port: 8081
+  servlet:
+    contextPath: /api
 
 cors:
   allowedOrigins:
